Add clear command
